把 view app 的 CSS 移到 resources/sass/app.scss，JS 移到 resources/js/app.js
